+++ app/src/main/java/com/masefal_0046/tigaunifess/ui/screen/MainViewModel.kt
@@ -18,7 +18,7 @@ class MainViewModel(private val fessRepository: FessRepository) : ViewModel() {
     )
 
     suspend fun getPesanById(id: Long): Pesan? {
-        return fessRepository.allPesan.first().find { it.id.toLong() == id }
+        return fessRepository.allPesan.first().find { it.id == id }
     }
 
     fun deletePesan(pesan: Pesan) {
